feat: Use progress now updating proving screen
